[oh-tab-zhwi.3.3] Extract HAL voice-confirm capture hook
